pyproject.toml
[project]
name = "knowledge-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "pymssql (>=2.3.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "azure-monitor-opentelemetry-exporter (>=1.0.0b36,<2.0.0)",
    "opentelemetry-instrumentation-fastapi (>=0.53b1,<0.54)",
    "opentelemetry-instrumentation-logging (>=0.53b1,<0.54)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
httpx = "^0.28.1"
pytest-cov = "^6.1.1"
flake8 = "^7.2.0"
black = "^25.1.0"
isort = "^6.0.1"

